refactor: move element-specific tests to element modules
- Move session tests from parser.rs to elements/sessions.rs
- Move list tests from parser.rs to elements/lists.rs
- Move definition tests from parser.rs to elements/definitions.rs
- Add Container trait imports where needed for label() method
- Clean up unused imports
- Keep integration tests (trifecta, ensemble) in parser.rs

All 181 tests pass (178 passed, 3 ignored). Tests are now co-located
with their respective element parsing logic.
