+++ src/txxt_nano/parser.rs
@@ -10,7 +10,6 @@
 
 pub mod api;
 pub mod combinators;
-pub mod conversion;
 pub mod document;
 pub mod elements;
 pub mod labels;
