--- src/txxt_nano/parser/conversion/mod.rs
@@ -1,6 +1,0 @@
-//! Conversion utilities for transforming intermediate AST to final AST
-//!
-//! This module contains functions for converting the intermediate AST (with spans)
-//! to the final AST (with extracted text) and position-aware variants.
-
-pub mod helpers;
